requirements.txt
beautifulsoup4~=4.10
requests~=2.22
google-api-python-client~=2.35
google-auth-httplib2~=0.1
google-auth-oauthlib~=0.4
